function based and class based views added

+++ store/function_based_views.py
@@ -7,87 +7,68 @@ from .models import Product, Collection
 from .serializers import ProductSerializer, CollectionSerializer
 from django.db.models.aggregates import Count
 from django.shortcuts import get_object_or_404
+
+
+# -----------------------------------------------------------------------------------------------------------
+# -----------------------------------------------------------------------------------------------------------
+# ------------------------------------------- FUNCTION BASED VIEWS ------------------------------------------
+# -----------------------------------------------------------------------------------------------------------
+# -----------------------------------------------------------------------------------------------------------
+
+
+
 # django way : 
 # def product_list(request):
 #     return HttpResponse('ok')
 
 # rest_framework way : 
 
-# @api_view(['GET', 'POST'])
-# def product_list(request):
+@api_view(['GET', 'POST'])
+def product_list(request):
 
-#     if request.method == 'GET':
-#         query_set = Product.objects.select_related('collection').all()
-#         serializer = ProductSerializer(query_set, many=True, context={'request':request})
-#         return Response(serializer.data)
-
-#     elif request.method =='POST':
-#         serializer = ProductSerializer(data=request.data)
-
-#         # if serializer.is_valid():
-#         #     serializer.validated_data
-#         #     return Response('ok')
-#         # else : 
-#         #     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-#         # MORE CLEAN WAY : 
-
-#         serializer.is_valid(raise_exception=True)
-#         serializer.save()
-#         return Response(serializer.data, status=status.HTTP_201_CREATED)
-
-class ProductList(APIView):
-    def get(self, request):
+    if request.method == 'GET':
         query_set = Product.objects.select_related('collection').all()
         serializer = ProductSerializer(query_set, many=True, context={'request':request})
         return Response(serializer.data)
 
-    def post(self, request):
+    elif request.method =='POST':
         serializer = ProductSerializer(data=request.data)
+
+        # if serializer.is_valid():
+        #     serializer.validated_data
+        #     return Response('ok')
+        # else : 
+        #     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+        # MORE CLEAN WAY : 
+
         serializer.is_valid(raise_exception=True)
         serializer.save()
         return Response(serializer.data, status=status.HTTP_201_CREATED)
-           
-
-# @api_view(['GET', 'PUT', 'DELETE'])
-# def product_detail(request, pk):
-#     product = get_object_or_404(Product, pk=pk)
-#     if request.method == 'GET': 
-#         # product = Product.objects.get(pk=id)
-#         serializer = ProductSerializer(product)
-#         return Response(serializer.data)
-
-#     elif request.method == 'PUT':
-#         serializer = ProductSerializer(product, data=request.data)
-#         serializer.is_valid(raise_exception=True)
-#         serializer.save()
-#         return Response(serializer.data)
-
-#     elif request.method == 'DELETE':
-#         product.delete()
-#         return Response(status=status.HTTP_204_NO_CONTENT)
 
 
-# class based product_detail :
 
-class ProductDetail(APIView):
-
-    def get(self, request,pk):
-        product = get_object_or_404(Product, pk=pk)
+@api_view(['GET', 'PUT', 'DELETE'])
+def product_detail(request, pk):
+    product = get_object_or_404(Product, pk=pk)
+    if request.method == 'GET': 
+        # product = Product.objects.get(pk=id)
         serializer = ProductSerializer(product)
         return Response(serializer.data)
 
-    def post(self, request, pk):
-        product = get_object_or_404(Product, pk=pk)
+    elif request.method == 'PUT':
         serializer = ProductSerializer(product, data=request.data)
         serializer.is_valid(raise_exception=True)
         serializer.save()
         return Response(serializer.data)
 
-    def delete(self, request, pk):
-        product = get_object_or_404(Product, pk=pk)
+    elif request.method == 'DELETE':
         product.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
+
+
+
+
